Header login logout interaction

+++ src/app/_shared/services/storage/storage.service.ts
@@ -1,38 +1,45 @@
-import { Injectable } from '@angular/core';
+import { Injectable, Inject, PLATFORM_ID } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
+import { BehaviorSubject } from 'rxjs';
 
 const USER_KEY = 'auth-user';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class StorageService {
+  private loggedInStatus = new BehaviorSubject<boolean>(this.isLoggedIn());
 
-  constructor() { }
+  constructor(@Inject(PLATFORM_ID) private platformId: Object) { }
 
   clean(): void {
-    window.sessionStorage.clear();
+    if (isPlatformBrowser(this.platformId)) {
+      window.sessionStorage.clear();
+      this.updateLoginStatus();
+    }
   }
 
   public saveUser(user: any): void {
-    window.sessionStorage.removeItem(USER_KEY);
-    window.sessionStorage.setItem(USER_KEY, JSON.stringify(user));
+    if (isPlatformBrowser(this.platformId)) {
+      window.sessionStorage.removeItem(USER_KEY);
+      window.sessionStorage.setItem(USER_KEY, JSON.stringify(user));
+      this.updateLoginStatus();
+    }
   }
 
-  public getUser(): any {
-    const user = window.sessionStorage.getItem(USER_KEY);
-    if (user) {
-      return JSON.parse(user);
-    }
-
-    return {};
+  private updateLoginStatus(): void {
+    this.loggedInStatus.next(this.isLoggedIn());
   }
 
   public isLoggedIn(): boolean {
-    const user = window.sessionStorage.getItem(USER_KEY);
-    if (user) {
-      return true;
+    if (isPlatformBrowser(this.platformId)) {
+      const user = window.sessionStorage.getItem(USER_KEY);
+      return !!user;
     }
-
     return false;
+  }
+
+  public getLoggedInStatus() {
+    return this.loggedInStatus.asObservable();
   }
 }
